single thread restore
