+++ mad_engine/examples/test2.rs
@@ -18,7 +18,7 @@ fn main() {
 }
 
 async fn test2_helper(name: &str) -> std::result::Result<(), EngineError> {
-    let handle = MadEngineHandle::new("data", name).await.unwrap();
+    let mut handle = MadEngineHandle::new("data", name).await.unwrap();
     handle.create("file2".to_string()).unwrap();
     info!("create file2 succeed...");
     let mut buf: Vec<u8> = vec![0u8; DATA_LEN];
