+++ mad_engine/src/lib.rs
@@ -2,6 +2,8 @@ pub mod common;
 pub use common::*;
 pub mod error;
 pub use error::*;
+pub mod device_engine;
+pub use device_engine::*;
 pub mod engine;
 pub use engine::*;
